chore: move everything to a single plugin group
