+++ assets/textures/car1/car1.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:41bfa318361c1ce51391fd42295b43858232efa9a3d6a7247cc680e2a5ac5cf3
+size 162678
