+++ assets/textures/car1/spikes1.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a00714c7d218715174d8234be8c880acb1fd4519c19902060145e9f07a83efa9
+size 26917
