+++ assets/textures/car1/wheel1.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a5841d60d6a25c565c0be715cf95139e6a57770e9e12a18db0aae113e444248b
+size 7792
