+++ assets/textures/car1/wheel2.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:add7cc2acd0857ee34c23592b2fb30708b3aedace3c1efe9ccdeec5268ba2067
+size 4755
